♻️ 複数のコンポーネントで使用するType文を切り離した
